Optimized a lot of code
