+++ src/Core/src/Jobs/BaseJob.php
@@ -19,9 +19,9 @@ namespace CashierProvider\Core\Jobs;
 
 use CashierProvider\Core\Billable;
 use CashierProvider\Core\Concerns\Config\Queue;
-use CashierProvider\Core\Data\Http\ResponseData;
 use CashierProvider\Core\Enums\RateLimiterEnum;
 use CashierProvider\Core\Exceptions\External\EmptyResponseException;
+use CashierProvider\Core\Http\Response;
 use CashierProvider\Core\Services\Driver;
 use Illuminate\Bus\Queueable;
 use Illuminate\Contracts\Queue\ShouldBeUnique;
@@ -46,7 +46,7 @@ abstract class BaseJob implements ShouldBeUnique, ShouldQueue
 
     public int $maxExceptions = 3;
 
-    abstract protected function request(): ResponseData;
+    abstract protected function request(): Response;
 
     public function __construct(
         public Model $payment,
@@ -69,7 +69,7 @@ abstract class BaseJob implements ShouldBeUnique, ShouldQueue
         $data = [
             'external_id'  => $response->getExternalId(),
             'operation_id' => $response->getOperationId(),
-            'info'         => $response->toArray(),
+            'info'         => $response,
         ];
 
         $this->payment->cashier
